style: run deno fmt to unify styling
The following patch updates js styling to use deno formatter[^1]. In
doing so, allowing to use a consistent styling. Removing the
need to fiddle with settings of what styling is used. Prettier[^2] is
another option that can be used to reach the same consensus on styling.

In this context, used Deno's formatter as it really gets out of the way
on formatting and is compatible with the style prettier would provide,
if using the same sane defaults.

The main prupose here was to clear out styling from upcoming patch for
[403][3] with actual changes the patch provides. Making it easier for
review what has changed.

[^1]: https://deno.land/manual/tools/formatter
[^2]: https://prettier.io/
[3]: https://github.com/bonfire-networks/bonfire-app/issues/403